refactored bookings, started on appl'n fee
