+++ commercialoperator/components/bookings/confirmation_pdf.py
@@ -179,178 +179,8 @@ class Remittance(Flowable):
             self.__payment_line()
         self.__footer_line()
 
-#def _set_template_group(mooring_var):
-
-
-#def get_template_group(mooring_var):
-
-
 
 def _create_header(canvas, doc, draw_page_number=True):
-    canvas.saveState()
-    canvas.setTitle('Invoice')
-    canvas.setFont(BOLD_FONTNAME, LARGE_FONTSIZE)
-
-    current_y = PAGE_HEIGHT - HEADER_MARGIN
-
-    dpaw_header_logo = ImageReader(DPAW_HEADER_LOGO)
-    dpaw_header_logo_size = dpaw_header_logo.getSize()
-    canvas.drawImage(dpaw_header_logo, PAGE_WIDTH / 3, current_y - (dpaw_header_logo_size[1]/2),width=dpaw_header_logo_size[0]/2, height=dpaw_header_logo_size[1]/2, mask='auto')
-
-    current_y -= 70
-    canvas.drawCentredString(PAGE_WIDTH / 2, current_y - LARGE_FONTSIZE, 'TAX INVOICE')
-
-    current_y -= 20
-    canvas.drawCentredString(PAGE_WIDTH / 2, current_y - LARGE_FONTSIZE, 'ABN: 38 052 249 024')
-
-    # Invoice address details
-    invoice_details_offset = 37
-    current_y -= 20
-    invoice = doc.invoice
-    canvas.setFont(BOLD_FONTNAME, SMALL_FONTSIZE)
-    current_x = PAGE_MARGIN + 5
-    canvas.drawString(current_x, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER),invoice.owner.get_full_name())
-    canvas.drawString(current_x, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 2,invoice.owner.username)
-    current_x += 452
-
-    #write Invoice details
-    canvas.drawString(current_x, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER),'Date')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER),invoice.created.strftime(DATE_FORMAT))
-    canvas.drawString(current_x, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 2, 'Page')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 2, str(canvas.getPageNumber()))
-    canvas.drawRightString(current_x + 20, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 3, 'Invoice Number')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 3, invoice.reference)
-    canvas.drawRightString(current_x + 20, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 4, 'Total (AUD)')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 4, currency(invoice.amount))
-    canvas.drawRightString(current_x + 20, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 5, 'GST included (AUD)')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 5, currency(invoice.amount - calculate_excl_gst(invoice.amount)))
-    canvas.drawRightString(current_x + 20, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 6, 'Paid (AUD)')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 6, currency(invoice.payment_amount))
-    canvas.drawRightString(current_x + 20, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 7, 'Outstanding (AUD)')
-    canvas.drawString(current_x + invoice_details_offset, current_y - (SMALL_FONTSIZE + HEADER_SMALL_BUFFER) * 7, currency(invoice.balance))
-    canvas.restoreState()
-
-def _create_invoice(invoice_buffer, invoice):
-
-    global DPAW_HEADER_LOGO
-#    if  cols_var["TEMPLATE_GROUP"] == 'rottnest':
-#        DPAW_HEADER_LOGO = os.path.join(settings.BASE_DIR, 'mooring', 'static', 'mooring', 'img','logo-rottnest-island-sm.png')
-#    else:
-#        DPAW_HEADER_LOGO = os.path.join(settings.BASE_DIR, 'ledger', 'payments','static', 'payments', 'img','dbca_logo.jpg')
-    DPAW_HEADER_LOGO = os.path.join(settings.BASE_DIR, 'ledger', 'payments','static', 'payments', 'img','dbca_logo.jpg')
-
-    every_page_frame = Frame(PAGE_MARGIN, PAGE_MARGIN + 250, PAGE_WIDTH - 2 * PAGE_MARGIN,
-                             PAGE_HEIGHT -450 , id='EveryPagesFrame',showBoundary=0)
-    remit_frame = Frame(PAGE_MARGIN, PAGE_MARGIN, PAGE_WIDTH - 2 * PAGE_MARGIN,
-                             PAGE_HEIGHT - 600, id='RemitFrame',showBoundary=0)
-    every_page_template = PageTemplate(id='EveryPages', frames=[every_page_frame,remit_frame], onPage=_create_header)
-
-
-    doc = BaseDocTemplate(invoice_buffer, pageTemplates=[every_page_template], pagesize=A4)
-
-
-    # this is the only way to get data into the onPage callback function
-    doc.invoice = invoice
-    owner = invoice.owner
-
-    elements = []
-    #elements.append(Spacer(1, SECTION_BUFFER_HEIGHT * 5))
-
-    # Draw Products Table
-    invoice_table_style = TableStyle([
-        ('VALIGN', (0, 0), (-1, -1), 'TOP'),
-        ('GRID',(0, 0), (-1, -1),1, colors.black),
-        ('ALIGN', (0, 0), (-1, -1), 'LEFT')
-        ])
-    items = invoice.order.lines.all()
-    discounts = invoice.order.basket_discounts
-    if invoice.text:
-        elements.append(Paragraph(invoice.text, styles['Left']))
-        elements.append(Spacer(1, SECTION_BUFFER_HEIGHT * 2))
-    data = [
-        ['Item','Product', 'Quantity','Unit Price', 'Total']
-    ]
-    val = 1
-    s = styles["BodyText"]
-    s.wordWrap = 'CJK'
-
-    for item in items:
-        data.append(
-            [
-                val,
-                Paragraph(item.description, s),
-                item.quantity,
-                currency(item.unit_price_incl_tax),
-                currency(item.line_price_before_discounts_incl_tax)
-            ]
-        )
-        val += 1
-    # Discounts
-    data.append(
-        [
-            '',
-            '',
-            '',
-            ''
-        ]
-    )
-    for discount in discounts:
-        data.append(
-            [
-                '',
-                discount.offer,
-                '',
-                '',
-                '-${}'.format(discount.amount)
-            ]
-        )
-        val += 1
-    t= Table(
-            data,
-            style=invoice_table_style,
-            hAlign='LEFT',
-            colWidths=(
-            0.7 * inch,
-            None,
-            0.7 * inch,
-            1.0 * inch,
-            1.0 * inch,
-            )
-        )
-    elements.append(t)
-    elements.append(Spacer(1, SECTION_BUFFER_HEIGHT * 2))
-    # /Products Table
-    if invoice.payment_status != 'paid' and invoice.payment_status != 'over_paid':
-        elements.append(Paragraph(settings.INVOICE_UNPAID_WARNING, styles['Left']))
-
-    elements.append(Spacer(1, SECTION_BUFFER_HEIGHT * 6))
-
-    # Remitttance Frame
-    elements.append(FrameBreak())
-    boundary = BrokenLine(PAGE_WIDTH - 2 * (PAGE_MARGIN *1.1))
-    elements.append(boundary)
-    elements.append(Spacer(1, SECTION_BUFFER_HEIGHT))
-
-    remittance = Remittance(HEADER_MARGIN,HEADER_MARGIN - 10,invoice)
-    elements.append(remittance)
-    #_create_remittance(invoice_buffer,doc)
-    doc.build(elements)
-
-    return invoice_buffer
-
-def create_invoice_pdf_bytes(filename, invoice):
-    invoice_buffer = BytesIO()
-    _create_invoice(invoice_buffer, invoice)
-
-    # Get the value of the BytesIO buffer
-    value = invoice_buffer.getvalue()
-    invoice_buffer.close()
-
-    return value
-
-# --------------------------------------------------------------------------------------------------------------
-
-def _create_confirmation_header(canvas, doc, draw_page_number=True):
     canvas.saveState()
     canvas.setTitle('Invoice')
     canvas.setFont(BOLD_FONTNAME, LARGE_FONTSIZE)
@@ -402,7 +232,7 @@ def _create_confirmation(confirmation_buffer, invoice, booking):
                              PAGE_HEIGHT -450 , id='EveryPagesFrame',showBoundary=0)
     remit_frame = Frame(PAGE_MARGIN, PAGE_MARGIN, PAGE_WIDTH - 2 * PAGE_MARGIN,
                              PAGE_HEIGHT - 600, id='RemitFrame',showBoundary=0)
-    every_page_template = PageTemplate(id='EveryPages', frames=[every_page_frame,remit_frame], onPage=_create_confirmation_header)
+    every_page_template = PageTemplate(id='EveryPages', frames=[every_page_frame,remit_frame], onPage=_create_header)
 
     doc = BaseDocTemplate(confirmation_buffer, pageTemplates=[every_page_template], pagesize=A4)
