+++ commercialoperator/components/main/utils.py
@@ -6,8 +6,7 @@ from django.core.cache import cache
 
 def retrieve_department_users():
     try:
-       # import ipdb; ipdb.set_trace()
-        res = requests.get('{}/api/users?minimal'.format(settings.CMS_URL), auth=(settings.LEDGER_USER,settings.LEDGER_PASS))
+        res = requests.get('{}/api/users?minimal'.format(settings.CMS_URL), auth=(settings.LEDGER_USER,settings.LEDGER_PASS), verify=False)
         res.raise_for_status()
         cache.set('department_users',json.loads(res.content).get('objects'),10800)
     except:
@@ -15,7 +14,7 @@ def retrieve_department_users():
 
 def get_department_user(email):
     try:
-        res = requests.get('{}/api/users?email={}'.format(settings.CMS_URL,email), auth=(settings.LEDGER_USER,settings.LEDGER_PASS))
+        res = requests.get('{}/api/users?email={}'.format(settings.CMS_URL,email), auth=(settings.LEDGER_USER,settings.LEDGER_PASS), verify=False)
         res.raise_for_status()
         data = json.loads(res.content).get('objects')
         if len(data) > 0:
